extra/protrade/requirements.txt
# Test requirements
pytest-django
model-bakery
faker
# Project requirements
django
pillow
ipython
